refactor and render menus
