+++ emurgo-haskell-project.cabal
@@ -24,7 +24,10 @@ library
     Sifo
 
   -- Modules included in this library but not exported.
-  -- other-modules:
+  other-modules:
+    Entry
+    Menu
+    Navigation
 
   -- LANGUAGE extensions used by modules in this package.
   -- other-extensions:
@@ -34,7 +37,9 @@ library
     , mtl
     , random
 
-  hs-source-dirs:   lib
+  hs-source-dirs:   lib lib/menus
+
+  --lib/menus
   default-language: Haskell2010
 
 executable emurgo-haskell-project
