+++ lib/menus/Navigation.hs
@@ -1,0 +1,9 @@
+module Navigation where
+
+type NavigationId = Int
+
+data Navigation
+  = Forward
+      { fOption :: NavigationId
+      }
+  | Back
